Changed paths: M       CHANGES.txt M       cat_conf_extopts.m M       cat_defaults.m M       cat_main.m M       cat_main_LAS.m M       cat_run_job.m A       cat_run_job1070.m M       cat_run_job_APP_SPM.m A       cat_run_job_APP_init1070.m M       cat_run_newcatch.m M       cat_vol_approx.m
(1) Changed:  Added "cat_run_job1070" and "cat_run_job_APP_init1070", update the
              GUI and set it as default APP preprocessing.
(2) Changed:  Removed last changes in "cat_main_LAS" of calling "cat_vol_approx"
              with automatic smoothness constrain that did not work yet. 


git-svn-id: svn://141.35.69.214/cat12@1145 9f44b166-36c4-483c-8295-4af200601734

+++ cat_vol_approx.m
@@ -89,8 +89,8 @@ function TA=cat_vol_approx(T,method,vx_vol,res,opt)
       TAr  = TAr .* (BMr | Tr);
       [MDr,MIr]  = cat_vbdist(single(TAr>0),TAr==0,double(resTr.vx_volr)); 
       TAr=TAr(MIr); TASr=cat_vol_smooth3X(TAr,4); TAr(~BMr)=TASr(~BMr);  clear TASr; 
-      TAr = cat_vol_laplace3R(TAr,~BMr,opt.lfO); TAr = cat_vol_median3(TAr,~BMr);
-      TAr = cat_vol_laplace3R(TAr,~Tr,opt.lfO); 
+      TAr = cat_vol_laplace3R(TAr,~BMr,double(opt.lfO)); TAr = cat_vol_median3(TAr,~BMr);
+      TAr = cat_vol_laplace3R(TAr,~Tr,double(opt.lfO)); 
     case 'linear'
       TNr = TAr;
       Tr  = TAr .* BMr;
